rpm: fix RPM_ETCCONFIGDIR value in SDK
The RPM_ETCCONFIGDIR should be some root directory instead of the
etc directory as in patch 0002-Add-support-for-prefixing-etc-from-RPM_ETCCONFIGDIR-.patch
RPM_ETCCONFIGDIR is used to prefix /etc.

The regression was introduced in the following patch.
"""
nativesdk-rpm: export RPM_ETCCONFIGDIR and MAGIC in environment like RPM_CONFIGDIR
"""
The patch incorrectly set RPM_ETCCONFIGDIR, maybe because this variable's
name is a little misleading.

Note that this patch modifies both RPM_ETCCONFIGDIR and RPM_CONFIGDIR.
The RPM_CONFIGDIR's value happened to be correct because of the ${libdir},
not $OECORE_NATIVE_SYSROOT. In fact, due to the substitution mechanism,
the $OECORE_NATIVE_SYSROOT is empty.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>