quilt: fix non-deterministic ownership in ptest package
The quilt package files are owned by the ptest user, which is provided
by the ptest-runner package.  The quilt recipe depends on ptest-runner
but there have been instances where rebuilds of quilt end up without the
right passwd being present, so the files are not owned by ptest.

I suspect what's happening is that the re-package is happening without
the ptest-runner package being present (reproducible with bitbake quilt;
bitbake quilt -cclean; bitbake quilt -Cpackage_write_rpm), so add an
explicit dependency on ptest-runner in the packaging dependencies to
ensure this doesn't happen.

[ YOCTO #15062  ]

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>