python3-psutil: fix-up -tests runtime dependencies
The psutil.tests module makes use of gcc, other tools and a minimal
set of headers and build tools. If these are missing you will see

======================================================================
ERROR: psutil.tests.test_process.TestProcess.test_weird_environ
----------------------------------------------------------------------
Traceback (most recent call last):
 File "./psutil/tests/test_process.py", line 1452, in test_weird_environ
   create_exe(path, c_code=code)
 File "./psutil/tests/__init__.py", line 831, in create_exe
   raise ValueError("gcc is not installed")
ValueError: gcc is not installed

along with other errors related to the missing tools and headers.

Add the missing RDEPENDS for the -tests package such that simply
adding the -tests package to an image will be all that is needed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>