defaultsetup: Enable largefile and 64bit time_t support systemwide for 32 bit platforms
There is a problem with 32 bit time looming for many 32 bit platforms
including arm, mip32, powerpc and x86 (but not riscv32) in 2038 when the
32 bit field time_t can no longer store the time value correctly due
to overflow.

Preparing for this is tricky as the ABI between libraries and binaries
will change and it isn't possible to migrate easily as structures and
return values change size.

As we're a source based system, the project has taken the decision that
we should change the compiler flags to switch to both largefile and 64
bit time_t. With OE-Core we've ironed out the issues we could spot
apart from some testing issues in strace and lttng-tools for which
discussions are ongoing upstream. There is more testing to be done
but we wanted to make this switch now in good time before our next
LTS release so we can work through any issues arrising.

We had already tried to mandate largefile everywhere before this but
this gives an opportunity to ensure that at the same time.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>