nettle: rewrite ptest integration
Instead of patching a relative path to an unversioned libnettle.so to be
a bare filename which then needs nettle-dev to be installed, create a
symlink in the expected place which points to the actual library. This
means nettle-ptest no longer needs to depend on nettle-dev.

Explicitly skip symbols-test, it has been silently failing as nm isn't
available and also needs a static libnettle.a to run.

Install two rsa-* example binaries that are needed for pkcs1-conv-test
to pass.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>