psplash: replace Yocto .h by .png splashscreen
This makes the image easier to modify.

People may not realize they can pass a .png image
to SPLASH_IMAGES, and producing a .h file is more
complicated.

Also provide the source SVG file (though the recipe doesn´t
use it, for easier modification) and support for higher resolutions
than the initial 640x480 image. This SVG file was reconstituted manually
to reproduce what was shown on the qemu86-64 image.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>