ptest-runner: Ensure data writes don't race
Currently the code can race as there is a read/write thread handling the stdio but
there is no guarantee that when the process exits, the thread has handled all the
data. This results in output where "END:" isn't actually at the end of the logs
but somewhere in the middle of the output.

Update to a revision with this fix.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(cherry picked from commit 6608ee551ed7e8864cdeabfb09832621313d3191)
Signed-off-by: Steve Sakoman <steve@sakoman.com>