sysfsutils: don't install to base_libdir
This library now ships a pkgconfig file but $base_libdir isn't in the
search path, so can't be used.

As the base_libdir/libdir split is pretty meaningless these days, simply
stop installing into base_libdir.  This was added to silence warnings
from pcmciautils[1] but PCMCIA isn't really a thing anymore so this
shouldn't be a problem.

[1] oe-core f3c93627

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>