weston: Cleanup and fix x11 and xwayland dependencies
For the x11 backend package config, drop the redundant dependencies
libxcb and cairo. The former is listed twice in the package config,
while the latter is also listed globally.

For the xwayland package config, add the missing dependencies libxcb
and libxcursor. These dependencies are hidden when the x11 backend is
enabled by default. When the default setting is overridden to disable
the x11 backend, the following error occurs:
```
| Run-time dependency xcursor found: NO (tried pkgconfig and cmake)
|
| ../git/xwayland/meson.build:34:2: ERROR: Problem encountered: xwayland requires xcursor which was not found. Or, you can use '-Dxwayland=false'.
```

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>