baremetal-helloworld: Fix race condition
There was a race condition during the Makefile execution in
between the assemble and compile targets, only the assemble
target had a dependency on creating the build directory.
If the compile target was executed first, an error was thrown
by bitbake since the build directory did not exist yet:

| Assembler messages:
| Fatal error: can't create build/hello_baremetal_aarch64.o:
  No such file or directory

Update the SRCREV to reflect the latest changes serializing
the makefile targets to avoid such race condition from happening.

[YOCTO #15146]

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>