linux-yocto/cve-exclusion: move entries from cve-extra-exclusions
We've a slew of CVE_CHECK_IGNOREs in cve-extra-exclusions which are to
mark a CVE as not valid with the current default kernel. However, this
file is kernel agnostic so if someone decides to build a 6.0 kernel then
these ignores are no longer valid.

Move the ignores which are to simply reflect backports to
cve-exclusions_6.1.inc so that they're version-specific. As the kernel
is upgraded these exclusions should be made redundant and removed from
the file.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>