glibc-testsuite: Fix network restrictions causing test failures
The check target for the glibc testsuite uses networking to access
a QEMU in system mode. This was resulting in large numbers of test
failures for the x86 targets that use it.

Enable network access to resolve this.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(cherry picked from commit 088e022f80e57ca345838bf57df9abad3d8b9313)
Signed-off-by: Steve Sakoman <steve@sakoman.com>